+++ application/models/Categories.php
@@ -1,0 +1,18 @@
+<?php
+class Categories extends CSV_Model
+{
+    // categories Id
+    public $id;
+    // categories name
+    public $name;
+    // categories description
+    public $description;
+
+    /**
+     * Contructor for Categories
+     */
+    function __construct()
+    {
+        parent::__construct("data/Categories.csv", "id", "entity");
+    }
+}