+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG.md
 
+### Note: Add changelog at the bottom of the file
+
 ## 1.0.0 (2018-02-08)
 
 Features:
@@ -9,3 +11,16 @@ Features:
       - 3 sections: items, slots, and stats.
   - JS
       - Listeners for drag-and-drop events.
+
+## 1.0.1 (2018-02-08)
+
+Features:
+
+    - Adding categories to the 'accessorizer' page
+    - Removed extraneous items. Sticking to 3 for each category.
+
+## 1.0.2 (2018-02-18)
+
+Features:
+
+    - Added categories model